Luce's homework (#8)
Jeremy established the core structure: decode the hex string, run it through a
decryptor whose key is each of every possible character, score the results,
order them by score, and pick the winner.

Right away Joe started us off with the idea to score maybe-decrypted messages
by the number of spaces (`<<32>>`). Given an arbitrary message length, we're
looking for the average length of words. The score is then the distance from
the average length of words in a modern English sentence: `4.7`. This would
turn out to be a very good and time-saving idea.

We were a bit off by XORing the entire message by one character. XORing each
byte by our cypher revealed the message somewhere in the brute-force output.
(This property also re-opens the possibility of working on a message of
arbitrary length or a stream.) It's possible that the challenge wanted us to
try every possible _char_ (`0..0x10FFFF`), but we started with every possible
byte (`0..255`) and that worked out.

I decided to curve and normalize the word length scores:
* curved – as a ratio of our-average-word-length to the
  English-average-word-length (rather than absolute distance), to help
  discriminate which results might be "interesting"; and with
* normalization - from 0 to 1, where 0 is a perfect score and 1 is "don't even
  consider me a contender"; this wasn't immediately necessary when I did it,
  but I anticipated needing more scoring modules and wanted to be able to
  weight them.

This bumped the English-looking result from somewhere in the top 20 results up
to number 2, which is great if you knew what you were looking for, but not good
enough for an algorithmic solution (as opposed to an assist).

I hunted around and found List.ascii_printabe?/1, added that to the scorer, and
weighted it as twice as useful as average word length. The desired result
buoyed to the top of the ranking with a confidence level of more than 18 times
the runner-up. :+1: